MD-494
Draw a dot line if the raw data overflows the displayed view

+++ WADODownload.m
@@ -360,7 +360,7 @@
             if( aborted)
                 NSLog( @"------ WADO downloading ABORTED");
             else
-                NSLog( @"------ WADO downloading : %d files - finished (errors: %d / total: %d)", (int) [urlToDownload count], countOfSuccesses, urlToDownload.count);
+                NSLog( @"------ WADO downloading : %d files - finished (errors: %d / total: %d)", (int) [urlToDownload count], countOfSuccesses, (int) urlToDownload.count);
         }
     }
     @catch (NSException *exception) {
